Add files via upload
This game is called Space Rocks. Of course, you probably already know that because it's written right in the title. There’s not much to it besides shooting rocks and scoring points. This is the first game I’ve created, inspired by the video https://youtu.be/nwlvT-L9vFg?si=uxxxfAYRxRtSKyqM, and it was made using Game Maker.